Move onboard into a separate package
